+++ src/User/User.php
@@ -74,6 +74,8 @@ final class User
     private $extra;
     /** @var \DateTimeImmutable|null */
     private $apiKeyUpdatedAt;
+    /** @var int|null */
+    private $passwordExpiryTimeLeft;
 
     /**
      * @internal
@@ -113,6 +115,7 @@ final class User
         $this->registeredAt = empty($data['registeredAt']) ? null : \DateTimeImmutable::createFromFormat(DATE_RFC3339, $data['registeredAt']);
         $this->extra = $data['extra'] ?? [];
         $this->apiKeyUpdatedAt = \array_key_exists('apiKeyUpdatedAt', $data) === true ? \DateTimeImmutable::createFromFormat(\DateTime::RFC3339, $data['apiKeyUpdatedAt']) : null;
+        $this->passwordExpiryTimeLeft = \array_key_exists('passwordExpiryTimeLeft', $data) === true ? \intval($data['passwordExpiryTimeLeft']) : null;
     }
 
     /**
@@ -342,5 +345,10 @@ final class User
     public function getApiKeyUpdatedAt(): ?\DateTimeImmutable
     {
         return $this->apiKeyUpdatedAt;
+    }
+
+    public function getPasswordExpiryTimeLeft(): ?int
+    {
+        return $this->passwordExpiryTimeLeft;
     }
 }